More styling * don't need animations * fix naming, maybe it still sux idk * add container for tabs
